refactor: 스토리북 decorators 파일 삭제 및 preview 파일에 로직 추가
