--- .storybook/preview.ts
@@ -1,16 +1,0 @@
-import type { Preview } from '@storybook/react';
-import withThemeProvider from './decorators';
-
-const preview: Preview = {
-  parameters: {
-    controls: {
-      matchers: {
-        color: /(background|color)$/i,
-        date: /Date$/i,
-      },
-    },
-  },
-  decorators: [withThemeProvider],
-};
-
-export default preview;
