+++ .storybook/tsconfig.json
@@ -1,0 +1,7 @@
+{
+  "extends": "../tsconfig.app.json",
+  "include": ["./**/*.ts", "./**/*.tsx"],
+  "compilerOptions": {
+    "jsx": "react-jsx"
+  }
+}
